+++ src/ewokscore/tests/examples/tasks/nooutputtask.py
@@ -1,6 +1,6 @@
 from pprint import pformat as _pformat  # hide from task discovery
 
-from ewokscore import Task
+from ....task import Task
 
 
 class NoOutputTask(Task):