+++ src/ewokscore/tests/test_method_task.py
@@ -1,6 +1,6 @@
 from ewoksutils.import_utils import qualname
 
-from ewokscore.task import Task
+from ..task import Task
 
 
 def mymethod1(a=0, b=0):