+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use `ewoksutils.uri_utils` and adapt to proper file URI's on Windows.
+  For example `file:///C:\Windows\Tmp`, note the leading `/` before `C:\`.
+
+### Removed
+
+- Module `ewokscore.persistence.uri`
+
 ## [3.1.0] - 2025-11-03
 
 ### Added